Adds foundational elements for main window and admin window
Most of the work has been done to the adminDubugMenu.ui for monday's showcase to the professor. The foundation is there but a lot of the implimentaion needs to still be done: needs to be connected to display list function, needs to have list sorter implimented and also needs to have add/remove connected to SQL which should be easy. The front end is functional enough for now.