printList function is implimented into the UI
Brandon's print list function is linked to the UI but is altered to output to a QListWidget instead. The function now takes in a QlistWidget pointer as a parameter and adds the resturant list to that QListWidget. Also debugMenu class was changed from a child of QMainWindow to a child of MainWindow, so that it could use the print function. MainWindow is already a child of QMainWindow...